science_agent/pyproject.toml
[project]
name = "science_agent"
version = "0.1.0"
description = "science_agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.175.0",
    "crewai-tools>=0.65.0",
    "gradio>=5.44.1",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rdkit>=2025.3.5",
]

[project.scripts]
science_agent = "science_agent.main:run"
run_crew = "science_agent.main:run"
train = "science_agent.main:train"
replay = "science_agent.main:replay"
test = "science_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
